+++ client/package.json
@@ -16,6 +16,8 @@
         "react-native-web": "^0.11.4",
         "react-redux": "^7.1.1",
         "redux": "^4.0.4",
+        "redux-saga": "^1.1.1",
+        "redux-saga-test-plan": "^4.0.0-rc.1",
         "typesafe-actions": "^4.4.2"
     },
     "devDependencies": {
@@ -24,6 +26,7 @@
         "@types/react-native": "^0.57.65",
         "@types/react-redux": "^7.1.2",
         "@types/redux": "^3.6.0",
+        "@types/redux-saga": "^0.10.5",
         "babel-preset-expo": "^6.0.0"
     }
 }
