+++ client/src/store/index.ts
@@ -2,44 +2,25 @@
  * This file is the entry point for our Redux store
  */
 
-import {
-    compose,
-    createStore as createReduxStore,
-    combineReducers,
-} from 'redux';
+import { compose, createStore as createReduxStore } from 'redux';
+import createSagaMiddleware from 'redux-saga';
 
-import {
-    demoReducer,
-    demoActions,
-    DemoAction,
-    demoSelectors,
-    IDemoState,
-} from './demo';
+import { reducers } from './reducers';
+export { actions, selectors, IRootState, Action } from './reducers';
 
-export interface IRootState {
-    demo: IDemoState;
-}
-
-export type Action = DemoAction;
-
-export const actions = {
-    demo: demoActions,
-};
-
-const reducers = combineReducers({
-    demo: demoReducer,
-});
-
-export const selectors = {
-    demo: demoSelectors,
-};
+import { initializeSagas } from './sagas';
 
 export const createStore = () => {
+    // Redux saga is used to introduce side-effects in our application
+    const reduxSaga = createSagaMiddleware();
+
     // We used the Redux chrome extension for debugging purposes
     const composeEnhancers =
         (window as any).__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
 
-    const store = createReduxStore(reducers, composeEnhancers());
+    const reduxStore = createReduxStore(reducers, composeEnhancers(reduxSaga));
 
-    return store;
+    reduxSaga.run(initializeSagas);
+
+    return reduxStore;
 };
