+++ client/src/store/sagas/demo/index.ts
@@ -1,0 +1,11 @@
+/**
+ * Demo saga; used for demonstration purposes
+ */
+
+import { put } from 'redux-saga/effects';
+
+import { actions } from '../..';
+
+export function* demoStart() {
+    yield put(actions.demo.changeDemoText('Hello from Saga land!'));
+}
