+++ client/src/store/sagas/index.test.ts
@@ -1,0 +1,12 @@
+import { expectSaga } from 'redux-saga-test-plan';
+
+import { initializeSagas } from '.';
+import { demoStart } from './demo';
+
+describe('initialize sagas', () => {
+    it('should start demo saga', () => {
+        return expectSaga(initializeSagas)
+            .spawn(demoStart)
+            .run();
+    });
+});
